Move view bill run feature into bill runs folder
