+++ cypress/integration/bill_runs/view/general/general_steps.js
@@ -1,5 +1,5 @@
 import { And, Then } from 'cypress-cucumber-preprocessor/steps'
-import TransactionEndpoints from '../../../endpoints/transaction_endpoints'
+import TransactionEndpoints from '../../../../endpoints/transaction_endpoints'
 
 And('I add a successful transaction with the following FY details', (dataTable) => {
   cy.wrap(dataTable.rawTable).each(row => {
